SCRD-1903 re-include kernel-firmware in autoyast packages
The kernel-firmware package was inadvertently dropped from the list of
packages to be installed by the SLES autoyast process when the SLES12
SP3 refresh was done. This patch adds the kernel-firmware package back
in.

Also restores the alphabetical sorting of the package list.

Ported-From: I13ba4ebc9e204561cb140ebdc47ddcae672e9065
(cherry-picked from https://gerrit.suse.provo.cloud/1749)

NOTE: For Ardana, the "porting" of the patches meant we re-cycled the
original SP2 version of the file and retained the kernel-firmware
reference.

Change-Id: Ia5d3f8682c190feac85a5a387b6b3439e904f5fd
Depends-On: I04871b62591bd64f4c2487c24b4f37ae8f1b28af