SCRD-1957 Fix run-cobbler CI breakage
Due to cherry-picking some port patches out of order the "re-add" of
the starting of SSH during the SLES12SP3 re-imaging was missing and
that caused us to fail the job as the node was uncontactable after
being re-imaged.

This patch re-adds the start of sshd, and cleans up the enablement
of services.

Change-Id: I04871b62591bd64f4c2487c24b4f37ae8f1b28af