SCRD-2251 Rebranding to ardana namespace (bsc#1081798)
RHEL anaconda kickstart config file still had hlm name references
for user home and user name variables. Those need to be changed to
ardana namespace so that expected values can be substituated in  its
template file.

Also fix hlmuser references in populate-rhel-per-profile.yml and a
hlm reference in get-sles-loader.yml.

Add wget to the list of packages installed so that we can use it
later in the kickstart scripts.

Change-Id: I0c2eb03020897dd5a37e55d368ef51b2e5e86958