Network check on Dl160 Gen9 during bm-reimage (bsc#1090751)
On certain DL160 Gen9 servers, bm-reimage run will fail at
TASK: "check the timestamp on the dest with local" because
the network on the client was not yet ready. This cause wget
call to get the preseed/timestamp file on the client to fail.
This fix adds a network check in the autoyast file to allow
the network to be ready before we call wget for the timestamp
file.

Change-Id: Id767694dd9f3ff9450b80d16e933e9165d3d8639