Simplify method to create grub2 files for UEFI boot (bsc#1093679)
This patch changes the mechanism for creating the grub2 files
needed for UEFI boot. The current method formats the grub
variable $net_defaul_mac (XX:XX:XX) using tr or regexp. This does
not work on servers with "Secure Boot" enabled as that mode
prevents grub from loading the tr or regexp modules. Instead,
these changes will just use the $net_default_variable which is
always available even in "Secure Boot" mode and create a file
with the same string (lowercase) in the name.

Change-Id: Ic088924317e69454e8fb5f39761588174fc9c752