Check for proper ipmi settings in servers.yml (bsc#1091459)
This patch checks to make sure there is at least one node
configured in servers.yml with ipmi credentials. If there are
not any nodes, cobbler-deploy exits. This is for the use case
where nodes may be provisioned outside of cobbler and not have ipmi
settings. In those scenarios, the cobbler-deploy playbook would
produce a non intuiive error message.  This patch adds a proper
message and stops the playbook from further processing.

Change-Id: I71742df408f6dcd3f8ec27998aaec6adfb80613d