SCRD-6821 gate cobbler import on distro existence (SCRD-6821)
The cobbler deploy was attempting to import SLES or RHEL distros
again, even if they already existed, based on the existence of
the <distro_ver>-x86_64 profile, rather than the existence of a
distro of that name; a profile of the same name will be created
automatically when a distro is initially imported but that can
be deleted without deleting the associated distro.

By ensuring that we check for the existence of the distro in
cobbler, rather than the profile, we avoid unnecessarily running
cobbler import again, which will fail.

Change-Id: I8fb071955b97f00d621f7cfa15ead8757b0529ad