SCRD-6821 Add retry handling around cobbler distro checks (SCRD-6821)
In some circumstances, cobbler commands are failing with an errno 111,
(Connection refused) error, so we need to add retry handling around
the checks for distro existence to ensure that we don't incorrectly
try to import a distro that already exists.

To handle this the cobbler check for a distro existing has been moved
to a separate task which checks for the specified distro existing,
wrapping the cobbler command in a retry loop.

Note that we may need to consider adding this retry handling around
more cobbler commands.

Change-Id: Idc4f4d3cec099499a99917427b79dfde77b077d2