Ensure distro_signatures.json gets updated if needed (SOC-11249)
The distro signatures files under /etc/cobbler and /var/lib/cobbler
are considered config files, and thus will not be replaced during an
update or upgrade of the cobbler package. This means that we may end
up with out of date versions of these files after update or upgrade,
and will need to update them as appropriate.

To determine if we need to take action we use rpm -V cobbler to see
if either of the distro signatures files are different with respect
to those packages in the rpm, and, if so we move them aside, with a
timestamp appended, and then force a re-install of the cobbler rpm.

Remove the inherited legacy code that was converting the installed
distro_signatures.json files from MSDOS to Unix end-of-line format
which was unnecessarily triggering cobbler to be re-installed even
if the signatures were up to date.

Change-Id: Ic218a20425e143fa14a28da004548a8e2f73624a